+++ src/builders/index.js
@@ -3,7 +3,7 @@ const HttpZError = require('../error')
 const RequestBuilder = require('./request')
 const ResponseBuilder = require('./response')
 
-module.exports = (messageModel) => {
+module.exports = (messageModel, opts = {}) => {
   if (_.isNil(messageModel)) {
     throw HttpZError.get('messageModel is required')
   }
@@ -11,7 +11,7 @@ module.exports = (messageModel) => {
     throw HttpZError.get('messageModel must be a plain object')
   }
   if (messageModel.method) {
-    return RequestBuilder.build(messageModel)
+    return RequestBuilder.build(messageModel, opts)
   }
   if (messageModel.statusCode) {
     return ResponseBuilder.build(messageModel)