+++ src/builders/request.js
@@ -1,18 +1,22 @@
+const _ = require('lodash')
 const consts = require('../consts')
 const validators = require('../validators')
+const utils = require('../utils')
+const HttpZError = require('../error')
 const Base = require('./base')
 
 class HttpZRequestBuilder extends Base {
-  static build(model) {
-    let instance = new HttpZRequestBuilder(model)
+  static build(...params) {
+    let instance = new HttpZRequestBuilder(...params)
     return instance.build()
   }
 
-  constructor({ method, protocolVersion, target, headers, body }) {
+  constructor({ method, protocolVersion, target, headers, body }, opts) {
     super({ headers, body })
     this.method = method
     this.protocolVersion = protocolVersion
     this.target = target
+    this.opts = opts
   }
 
   build() {
@@ -33,6 +37,18 @@ class HttpZRequestBuilder extends Base {
       this.target + ' ' +
       this.protocolVersion.toUpperCase() +
       consts.EOL
+  }
+
+  _generateHeaderRows() {
+    validators.validateArray(this.headers, 'headers')
+    if (this.opts.mandatoryHost) {
+      let hostHeader = _.find(this.headers, name => utils.pretifyHeaderName(name) === consts.http.headers.host)
+      if (!hostHeader) {
+        throw HttpZError.get('Host header is required')
+      }
+    }
+
+    return super._generateHeaderRows()
   }
 }